Refactor import statements to use consistent module paths for Ducobox sensor classes
